// test/integration/test/ports/engines/queries/filters/field_reference/failure/_fixture/default/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Child: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly str: CodecTypes['pg/text@1']['output'];
      readonly testId: CodecTypes['pg/int4@1']['output'];
    };
    readonly TestModel: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly str: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Child: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly str: CodecTypes['pg/text@1']['input'];
      readonly testId: CodecTypes['pg/int4@1']['input'];
    };
    readonly TestModel: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly str: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly child: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly str: CodecTypes['pg/text@1']['output'];
      readonly testId: CodecTypes['pg/int4@1']['output'];
    };
    readonly testModel: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly str: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly child: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly str: CodecTypes['pg/text@1']['input'];
      readonly testId: CodecTypes['pg/int4@1']['input'];
    };
    readonly testModel: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly str: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;